Part 1 Bonus

Create a list of all funnel words that exist in a dictionary.

Usage

pt1_bonus(x, search_func)

Arguments

x

The word used to create funnel words.

search_func

A function to check to see if a candidate word is in the a dictionary.

See also

Examples

if (FALSE) { # \dontrun{
data(enable1)
set_search_function <- set_search(enable1)
pt1_bonus("dragoon", set_search_function)
} # }